moved prompt to prompt_local
