+++ src/prompt_local/README.md
@@ -1,0 +1,1 @@
+Prompt from a model running locally on a GPU node
